Talk for ARL-HPC Center (Powerpoint slides)
This is a talk for the Army Research Lab High Performance Computing Center on ParaLanczos that was given on November 21, 2016.  These slides explain some of the theory and motivation for this library.  

----

ParaLanczos:  A Scalable, Parallel GPU-based Lanczos Algorithm in CUDA

Abstract:

The Lanczos algorithm is an important tool for tridiagonalization and eigendecompositions of symmetric matrices in numerical computation: it is implemented in libraries such as ARPACK and Intel’s Math Kernel Library, and is also used in such well-known user-end software packages as MATLAB.   However, most extant implementations are focused on large or sparse matrices; this is ill-suited for the field of compressive sensing, where one must check the eigenvalues of large quantities of relatively small, dense submatrices within a dense positive-symmetric matrix.  The ParaLanczos algorithm aims to close the gap, providing researchers with a versatile tool for studying sen